Revert "finish linux pipewire"
This reverts commit 1999892113688cbe89e8a2d5c012633de3cc0dc0.
