+++ server-core/CMakeLists.txt
@@ -6,29 +6,16 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
-if(WIN32)
+if (WIN32)
 	add_compile_definitions(_UNICODE UNICODE)
 endif()
 
-if(${CMAKE_SYSTEM_NAME} STREQUAL "Windows")
-	set(PLATFORM_NAME "win32")
-elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
-	set(PLATFORM_NAME "linux")
-endif()
-
-message(STATUS "PLATFORMA_NAME ${PLATFORM_NAME}")
-
-find_package(Protobuf CONFIG REQUIRED)
-include_directories(${CMAKE_CURRENT_BINARY_DIR})
-protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS ../protos/client.proto)
-
-include_directories("src")
 set(
 	lib_src_list
-	"src/network_manager.cpp"
+	"src/client.pb.cc"
 	"src/audio_manager.cpp"
-	"src/${PLATFORM_NAME}/audio_manager_impl.cpp"
-	${PROTO_SRCS}
+	"src/common.cpp"
+	"src/network_manager.cpp"
 )
 
 add_executable(server-cmd
@@ -48,7 +35,3 @@ target_link_libraries(server-cmd PRIVATE protobuf::libprotobuf)
 
 find_package(cxxopts CONFIG REQUIRED)
 target_link_libraries(server-cmd PRIVATE cxxopts::cxxopts)
-
-find_package(PkgConfig REQUIRED)
-pkg_check_modules(pipewire REQUIRED IMPORTED_TARGET libpipewire-0.3)
-target_link_libraries(server-cmd PRIVATE PkgConfig::pipewire)
